rename: style 폴더 styles 폴더로 이름 변경
