+++ apps/web/features/auction/ui/styles/AuctionCard.styles.ts
