+++ apps/web/features/auction/ui/styles/AuctionOverlay.ts
