+++ apps/web/widgets/auction-detail-card/ui/styles/AuctionDescriptionCard.styles.ts
