+++ apps/web/widgets/auction-detail-card/ui/styles/AuctionDetailCard.styles.ts
