+++ apps/web/widgets/auction-detail-card/ui/styles/AuctionSellerProfile.styles.ts
